refactor(invitations): split InvitationListTab under the line cap (#544)
Extract two cohesive seams: a shared generic InvitationTable.svelte (the
registered/pending tables, differing only in identity columns — passed as
typed snippets — plus the shared invitationProperties snippet and
formatDate), and InvitationFormDialogs.svelte (the create/edit/bulk-edit
dialogs and their two shared form snippets, exposing openCreate/openEdit/
openBulk via bind:this like the DiscountCodeInput precedent).

The SvelteSet selection state (#535 reactivity fix) stays declared in the
parent and is passed by reference; children mutate in place only. The
toggleSelectAll allSelected-before-clear ordering is preserved intact in
the table. Behavior-preserving extraction only.

1090 -> 213 lines (InvitationTable: 261, InvitationFormDialogs: 614).
