+++ src/lib/components/invitations/InvitationListTab.svelte
@@ -5,15 +5,13 @@
 		PendingEventInvitationListSchema,
 		TicketTierDetailSchema
 	} from '$lib/api/generated/types.gen';
-	import { enhance } from '$app/forms';
-	import { Search, Mail, UserPlus, Trash2, Plus, Edit, CheckSquare, Square } from '@lucide/svelte';
-	import { formatDistanceToNow } from 'date-fns';
+	import { Search, Mail, UserPlus, Plus, Edit } from '@lucide/svelte';
 	import { getUserDisplayName } from '$lib/utils/user-display';
 	import { Button } from '$lib/components/ui/button';
-	import * as Dialog from '$lib/components/ui/dialog';
 	import UserAvatar from '$lib/components/common/UserAvatar.svelte';
-	import EmailTagInput from '$lib/components/forms/EmailTagInput.svelte';
 	import { SvelteSet } from 'svelte/reactivity';
+	import InvitationTable from './InvitationTable.svelte';
+	import InvitationFormDialogs from './InvitationFormDialogs.svelte';
 
 	// Only totalCount is read here; callers pass wider pagination objects.
 	interface Pagination {
@@ -46,158 +44,25 @@
 
 	let processingId = $state<string | null>(null);
 
-	// Create invitation form state
-	let showCreateDialog = $state(false);
-	let invitationEmails = $state('');
-	let invitationMessage = $state('');
-	let emailTags = $state<string[]>([]);
-	let createTierIds = $state<string[]>([]);
-
-	// Edit invitation state
-	let showEditDialog = $state(false);
-	let editingInvitation = $state<
-		EventInvitationListSchema | PendingEventInvitationListSchema | null
-	>(null);
-	let editingType = $state<'registered' | 'pending' | null>(null);
-	let editFormData = $state({
-		waives_questionnaire: false,
-		waives_purchase: false,
-		waives_membership_required: false,
-		waives_rsvp_deadline: false,
-		overrides_max_attendees: false,
-		custom_message: '',
-		tier_ids: [] as string[]
-	});
-
-	// Bulk selection state
+	// Bulk selection state (SvelteSet keeps selection reactive — kept in the
+	// parent and passed by reference; child tables mutate it in place).
 	const selectedRegisteredIds = new SvelteSet<string>();
 	const selectedPendingIds = new SvelteSet<string>();
-	let showBulkEditDialog = $state(false);
-	let bulkEditFormData = $state({
-		waives_questionnaire: false,
-		waives_purchase: false,
-		waives_membership_required: false,
-		waives_rsvp_deadline: false,
-		overrides_max_attendees: false,
-		custom_message: '',
-		tier_ids: [] as string[]
-	});
 
 	const totalSelected = $derived(selectedRegisteredIds.size + selectedPendingIds.size);
 
-	function formatDate(dateString: string): string {
-		try {
-			const date = new Date(dateString);
-			return formatDistanceToNow(date, { addSuffix: true });
-		} catch {
-			return dateString;
-		}
-	}
-
-	function resetCreateForm() {
-		invitationEmails = '';
-		invitationMessage = '';
-		emailTags = [];
-		createTierIds = [];
-	}
-
-	function openEditDialog(
-		invitation: EventInvitationListSchema | PendingEventInvitationListSchema,
-		type: 'registered' | 'pending'
-	) {
-		editingInvitation = invitation;
-		editingType = type;
-		editFormData = {
-			waives_questionnaire: invitation.waives_questionnaire,
-			waives_purchase: invitation.waives_purchase,
-			waives_membership_required: invitation.waives_membership_required,
-			waives_rsvp_deadline: invitation.waives_rsvp_deadline,
-			overrides_max_attendees: invitation.overrides_max_attendees,
-			custom_message: invitation.custom_message || '',
-			tier_ids: invitation.tiers?.map((t) => t.id) ?? []
-		};
-		showEditDialog = true;
-	}
-
-	function resetEditForm() {
-		editingInvitation = null;
-		editingType = null;
-		editFormData = {
-			waives_questionnaire: false,
-			waives_purchase: false,
-			waives_membership_required: false,
-			waives_rsvp_deadline: false,
-			overrides_max_attendees: false,
-			custom_message: '',
-			tier_ids: []
-		};
-	}
-
-	function toggleRegisteredSelection(id: string) {
-		if (selectedRegisteredIds.has(id)) {
-			selectedRegisteredIds.delete(id);
-		} else {
-			selectedRegisteredIds.add(id);
-		}
-	}
-
-	function togglePendingSelection(id: string) {
-		if (selectedPendingIds.has(id)) {
-			selectedPendingIds.delete(id);
-		} else {
-			selectedPendingIds.add(id);
-		}
-	}
-
-	function toggleSelectAllRegistered() {
-		const allSelected = selectedRegisteredIds.size === registeredInvitations.length;
-		selectedRegisteredIds.clear();
-		if (!allSelected) {
-			for (const inv of registeredInvitations) selectedRegisteredIds.add(inv.id);
-		}
-	}
-
-	function toggleSelectAllPending() {
-		const allSelected = selectedPendingIds.size === pendingInvitations.length;
-		selectedPendingIds.clear();
-		if (!allSelected) {
-			for (const inv of pendingInvitations) selectedPendingIds.add(inv.id);
-		}
-	}
-
-	function openBulkEditDialog() {
-		bulkEditFormData = {
-			waives_questionnaire: false,
-			waives_purchase: false,
-			waives_membership_required: false,
-			waives_rsvp_deadline: false,
-			overrides_max_attendees: false,
-			custom_message: '',
-			tier_ids: []
-		};
-		showBulkEditDialog = true;
-	}
-
-	function resetBulkEditForm() {
-		bulkEditFormData = {
-			waives_questionnaire: false,
-			waives_purchase: false,
-			waives_membership_required: false,
-			waives_rsvp_deadline: false,
-			overrides_max_attendees: false,
-			custom_message: '',
-			tier_ids: []
-		};
-	}
+	let dialogs: InvitationFormDialogs | undefined = $state();
 
 	function clearSelections() {
 		selectedRegisteredIds.clear();
 		selectedPendingIds.clear();
 	}
 
-	function handleEmailTagsChange(tags: string[]) {
-		emailTags = tags;
-		invitationEmails = tags.join('\n');
+	function openEdit(
+		invitation: EventInvitationListSchema | PendingEventInvitationListSchema,
+		type: 'registered' | 'pending'
+	) {
+		dialogs?.openEdit(invitation, type);
 	}
 </script>
 
@@ -209,12 +74,12 @@
 		</p>
 		<div class="flex gap-2">
 			{#if totalSelected > 0}
-				<Button variant="outline" onclick={openBulkEditDialog}>
+				<Button variant="outline" onclick={() => dialogs?.openBulk()}>
 					<Edit class="h-4 w-4" aria-hidden="true" />
 					{m['eventInvitationsAdmin.editSelected']({ count: totalSelected })}
 				</Button>
 			{/if}
-			<Button onclick={() => (showCreateDialog = true)}>
+			<Button onclick={() => dialogs?.openCreate()}>
 				<Plus class="h-4 w-4" aria-hidden="true" />
 				{m['eventInvitationsAdmin.createInvitations']()}
 			</Button>
@@ -237,854 +102,112 @@
 	</div>
 
 	<!-- Registered Invitations -->
-	<div class="space-y-3">
-		<div class="flex items-center justify-between">
-			<h3 class="text-lg font-semibold">
-				{m['eventInvitationsAdmin.registeredUsersTitle']({
-					count: registeredPagination.totalCount
-				})}
-			</h3>
-			{#if selectedRegisteredIds.size > 0}
-				<div class="flex items-center gap-2">
-					<span class="text-sm text-muted-foreground">
-						{m['eventInvitationsAdmin.selected']({ count: selectedRegisteredIds.size })}
-					</span>
-					<Button size="sm" variant="outline" onclick={clearSelections}
-						>{m['eventInvitationsAdmin.clear']()}</Button
-					>
-				</div>
-			{/if}
-		</div>
-
-		{#if registeredInvitations.length === 0}
-			<div class="rounded-lg border bg-card p-8 text-center">
-				<UserPlus class="mx-auto mb-2 h-8 w-8 text-muted-foreground" aria-hidden="true" />
-				<p class="text-sm text-muted-foreground">
-					{m['eventInvitationsAdmin.noRegisteredInvitations']()}
-				</p>
-			</div>
-		{:else}
-			<div class="overflow-hidden rounded-lg border bg-card">
-				<div class="overflow-x-auto">
-					<table class="w-full">
-						<thead class="border-b bg-muted/50">
-							<tr>
-								<th class="w-12 px-4 py-3">
-									<button
-										type="button"
-										onclick={toggleSelectAllRegistered}
-										class="flex items-center justify-center text-muted-foreground hover:text-foreground"
-									>
-										{#if selectedRegisteredIds.size === registeredInvitations.length && registeredInvitations.length > 0}
-											<CheckSquare class="h-4 w-4" aria-hidden="true" />
-										{:else}
-											<Square class="h-4 w-4" aria-hidden="true" />
-										{/if}
-									</button>
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerUser']()}
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerEmail']()}
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerProperties']()}
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerCreated']()}
-								</th>
-								<th
-									class="px-4 py-3 text-right text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerActions']()}
-								</th>
-							</tr>
-						</thead>
-						<tbody class="divide-y">
-							{#each registeredInvitations as invitation (invitation.id)}
-								<tr class="transition-colors hover:bg-muted/50">
-									<!-- Checkbox -->
-									<td class="px-4 py-4">
-										<button
-											type="button"
-											onclick={() => toggleRegisteredSelection(invitation.id)}
-											class="flex items-center justify-center text-muted-foreground hover:text-foreground"
-										>
-											{#if selectedRegisteredIds.has(invitation.id)}
-												<CheckSquare class="h-4 w-4" aria-hidden="true" />
-											{:else}
-												<Square class="h-4 w-4" aria-hidden="true" />
-											{/if}
-										</button>
-									</td>
-
-									<!-- User -->
-									<td class="px-4 py-4">
-										<div class="flex items-center gap-2">
-											<UserAvatar
-												profilePictureUrl={invitation.user.profile_picture_url}
-												previewUrl={invitation.user.profile_picture_preview_url}
-												thumbnailUrl={invitation.user.profile_picture_thumbnail_url}
-												displayName={getUserDisplayName(
-													invitation.user,
-													m['eventInvitationsAdmin.unknownUser']()
-												)}
-												firstName={invitation.user.first_name}
-												lastName={invitation.user.last_name}
-												size="sm"
-												clickable={true}
-											/>
-											<span class="text-sm font-medium">
-												{getUserDisplayName(
-													invitation.user,
-													m['eventInvitationsAdmin.unknownUser']()
-												)}
-											</span>
-										</div>
-									</td>
-
-									<!-- Email -->
-									<td class="px-4 py-4 text-sm text-muted-foreground">
-										{invitation.user.email || 'N/A'}
-									</td>
-
-									<!-- Properties -->
-									<td class="px-4 py-4">
-										{@render invitationProperties(invitation)}
-									</td>
-
-									<!-- Created -->
-									<td class="px-4 py-4 text-sm text-muted-foreground">
-										{formatDate(invitation.created_at)}
-									</td>
-
-									<!-- Actions -->
-									<td class="px-4 py-4 text-right">
-										<div class="flex items-center justify-end gap-2">
-											<button
-												type="button"
-												onclick={() => openEditDialog(invitation, 'registered')}
-												class="inline-flex items-center gap-1 rounded-md bg-primary px-2 py-1 text-xs font-medium text-primary-foreground transition-colors hover:bg-primary/90"
-											>
-												<Edit class="h-3 w-3" aria-hidden="true" />
-												{m['eventInvitationsAdmin.edit']()}
-											</button>
-											<form
-												method="POST"
-												action="?/deleteInvitation"
-												use:enhance={() => {
-													processingId = invitation.id;
-													return async ({ update }) => {
-														await update();
-														processingId = null;
-														selectedRegisteredIds.delete(invitation.id);
-													};
-												}}
-											>
-												<input type="hidden" name="invitation_id" value={invitation.id} />
-												<input type="hidden" name="invitation_type" value="registered" />
-												<button
-													type="submit"
-													disabled={processingId === invitation.id}
-													class="inline-flex items-center gap-1 rounded-md bg-destructive px-2 py-1 text-xs font-medium text-destructive-foreground transition-colors hover:bg-destructive/90 disabled:cursor-not-allowed disabled:opacity-50"
-												>
-													<Trash2 class="h-3 w-3" aria-hidden="true" />
-													{m['eventInvitationsAdmin.delete']()}
-												</button>
-											</form>
-										</div>
-									</td>
-								</tr>
-							{/each}
-						</tbody>
-					</table>
-				</div>
-			</div>
-		{/if}
-	</div>
+	<InvitationTable
+		title={m['eventInvitationsAdmin.registeredUsersTitle']({
+			count: registeredPagination.totalCount
+		})}
+		invitations={registeredInvitations}
+		selectedIds={selectedRegisteredIds}
+		invitationType="registered"
+		bind:processingId
+		onClear={clearSelections}
+		onEdit={openEdit}
+		identityHeaders={registeredHeaders}
+		identityCells={registeredCells}
+		emptyState={registeredEmpty}
+	/>
 
 	<!-- Pending Invitations -->
-	<div class="space-y-3">
-		<div class="flex items-center justify-between">
-			<h3 class="text-lg font-semibold">
-				{m['eventInvitationsAdmin.pendingUsersTitle']({
-					count: pendingPagination.totalCount
-				})}
-			</h3>
-			{#if selectedPendingIds.size > 0}
-				<div class="flex items-center gap-2">
-					<span class="text-sm text-muted-foreground">
-						{m['eventInvitationsAdmin.selected']({ count: selectedPendingIds.size })}
-					</span>
-					<Button size="sm" variant="outline" onclick={clearSelections}
-						>{m['eventInvitationsAdmin.clear']()}</Button
-					>
-				</div>
-			{/if}
-		</div>
-
-		{#if pendingInvitations.length === 0}
-			<div class="rounded-lg border bg-card p-8 text-center">
-				<Mail class="mx-auto mb-2 h-8 w-8 text-muted-foreground" aria-hidden="true" />
-				<p class="text-sm text-muted-foreground">
-					{m['eventInvitationsAdmin.noPendingInvitations']()}
-				</p>
-			</div>
-		{:else}
-			<div class="overflow-hidden rounded-lg border bg-card">
-				<div class="overflow-x-auto">
-					<table class="w-full">
-						<thead class="border-b bg-muted/50">
-							<tr>
-								<th class="w-12 px-4 py-3">
-									<button
-										type="button"
-										onclick={toggleSelectAllPending}
-										class="flex items-center justify-center text-muted-foreground hover:text-foreground"
-									>
-										{#if selectedPendingIds.size === pendingInvitations.length && pendingInvitations.length > 0}
-											<CheckSquare class="h-4 w-4" aria-hidden="true" />
-										{:else}
-											<Square class="h-4 w-4" aria-hidden="true" />
-										{/if}
-									</button>
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerEmail']()}
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerProperties']()}
-								</th>
-								<th
-									class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerCreated']()}
-								</th>
-								<th
-									class="px-4 py-3 text-right text-xs font-medium uppercase tracking-wider text-muted-foreground"
-								>
-									{m['eventInvitationsAdmin.headerActions']()}
-								</th>
-							</tr>
-						</thead>
-						<tbody class="divide-y">
-							{#each pendingInvitations as invitation (invitation.id)}
-								<tr class="transition-colors hover:bg-muted/50">
-									<!-- Checkbox -->
-									<td class="px-4 py-4">
-										<button
-											type="button"
-											onclick={() => togglePendingSelection(invitation.id)}
-											class="flex items-center justify-center text-muted-foreground hover:text-foreground"
-										>
-											{#if selectedPendingIds.has(invitation.id)}
-												<CheckSquare class="h-4 w-4" aria-hidden="true" />
-											{:else}
-												<Square class="h-4 w-4" aria-hidden="true" />
-											{/if}
-										</button>
-									</td>
-
-									<!-- Email -->
-									<td class="px-4 py-4 text-sm font-medium">{invitation.email}</td>
-
-									<!-- Properties -->
-									<td class="px-4 py-4">
-										{@render invitationProperties(invitation)}
-									</td>
-
-									<!-- Created -->
-									<td class="px-4 py-4 text-sm text-muted-foreground">
-										{formatDate(invitation.created_at)}
-									</td>
-
-									<!-- Actions -->
-									<td class="px-4 py-4 text-right">
-										<div class="flex items-center justify-end gap-2">
-											<button
-												type="button"
-												onclick={() => openEditDialog(invitation, 'pending')}
-												class="inline-flex items-center gap-1 rounded-md bg-primary px-2 py-1 text-xs font-medium text-primary-foreground transition-colors hover:bg-primary/90"
-											>
-												<Edit class="h-3 w-3" aria-hidden="true" />
-												{m['eventInvitationsAdmin.edit']()}
-											</button>
-											<form
-												method="POST"
-												action="?/deleteInvitation"
-												use:enhance={() => {
-													processingId = invitation.id;
-													return async ({ update }) => {
-														await update();
-														processingId = null;
-														selectedPendingIds.delete(invitation.id);
-													};
-												}}
-											>
-												<input type="hidden" name="invitation_id" value={invitation.id} />
-												<input type="hidden" name="invitation_type" value="pending" />
-												<button
-													type="submit"
-													disabled={processingId === invitation.id}
-													class="inline-flex items-center gap-1 rounded-md bg-destructive px-2 py-1 text-xs font-medium text-destructive-foreground transition-colors hover:bg-destructive/90 disabled:cursor-not-allowed disabled:opacity-50"
-												>
-													<Trash2 class="h-3 w-3" aria-hidden="true" />
-													{m['eventInvitationsAdmin.delete']()}
-												</button>
-											</form>
-										</div>
-									</td>
-								</tr>
-							{/each}
-						</tbody>
-					</table>
-				</div>
-			</div>
-		{/if}
-	</div>
+	<InvitationTable
+		title={m['eventInvitationsAdmin.pendingUsersTitle']({
+			count: pendingPagination.totalCount
+		})}
+		invitations={pendingInvitations}
+		selectedIds={selectedPendingIds}
+		invitationType="pending"
+		bind:processingId
+		onClear={clearSelections}
+		onEdit={openEdit}
+		identityHeaders={pendingHeaders}
+		identityCells={pendingCells}
+		emptyState={pendingEmpty}
+	/>
 </div>
 
-<!-- Create Invitation Dialog -->
-<Dialog.Root open={showCreateDialog} onOpenChange={(open) => (showCreateDialog = open)}>
-	<Dialog.Content class="flex max-h-[90dvh] flex-col sm:max-w-[600px]">
-		<Dialog.Header>
-			<Dialog.Title>{m['eventInvitationsAdmin.createInvitations']()}</Dialog.Title>
-			<Dialog.Description>
-				{m['eventInvitationsAdmin.createDialogDescription']()}
-			</Dialog.Description>
-		</Dialog.Header>
+<InvitationFormDialogs
+	bind:this={dialogs}
+	{organizationSlug}
+	{accessToken}
+	{registeredInvitations}
+	{pendingInvitations}
+	{selectedRegisteredIds}
+	{selectedPendingIds}
+	{ticketTiers}
+	onClearSelections={clearSelections}
+/>
 
-		<form
-			method="POST"
-			action="?/createInvitations"
-			use:enhance={() => {
-				return async ({ update }) => {
-					await update();
-					resetCreateForm();
-					showCreateDialog = false;
-				};
-			}}
-			class="flex min-h-0 flex-1 flex-col gap-4"
-		>
-			<div class="min-h-0 flex-1 space-y-4 overflow-y-auto pr-1">
-				<!-- Hidden inputs for form submission -->
-				<input type="hidden" name="emails" value={invitationEmails} />
-				<input type="hidden" name="tier_ids" value={JSON.stringify(createTierIds)} />
-
-				<!-- Email addresses with tag input -->
-				<EmailTagInput
-					{emailTags}
-					{organizationSlug}
-					{accessToken}
-					placeholder={m['eventInvitationsAdmin.emailPlaceholder']()}
-					onTagsChange={handleEmailTagsChange}
-				/>
-
-				<!-- Custom message -->
-				<div>
-					<label for="custom_message" class="block text-sm font-medium">
-						{m['eventInvitationsAdmin.customMessageLabel']()}
-					</label>
-					<textarea
-						id="custom_message"
-						name="custom_message"
-						bind:value={invitationMessage}
-						placeholder={m['eventInvitationsAdmin.customMessagePlaceholder']()}
-						rows="3"
-						maxlength="500"
-						class="mt-1 w-full rounded-md border-2 border-gray-300 bg-white px-3 py-2 text-sm transition-colors placeholder:text-muted-foreground focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary focus:ring-offset-2 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100"
-					></textarea>
-					<p class="mt-1 text-xs text-muted-foreground">
-						{m['eventInvitationsAdmin.charactersCount']({ count: invitationMessage.length })}
-					</p>
-				</div>
-
-				<!-- Invitation properties -->
-				{@render invitationPropertiesCheckboxes()}
-
-				<!-- Ticket tier selection -->
-				{@render tierCheckboxes(createTierIds, (ids) => (createTierIds = ids))}
-			</div>
-
-			<Dialog.Footer>
-				<Button type="button" variant="outline" onclick={() => (showCreateDialog = false)}>
-					{m['eventInvitationsAdmin.cancel']()}
-				</Button>
-				<Button type="submit">
-					<Mail class="mr-2 h-4 w-4" aria-hidden="true" />
-					{m['eventInvitationsAdmin.sendInvitations']()}
-				</Button>
-			</Dialog.Footer>
-		</form>
-	</Dialog.Content>
-</Dialog.Root>
-
-<!-- Edit Invitation Dialog -->
-<Dialog.Root
-	open={showEditDialog}
-	onOpenChange={(open) => {
-		showEditDialog = open;
-		if (!open) resetEditForm();
-	}}
->
-	<Dialog.Content class="flex max-h-[90dvh] flex-col sm:max-w-[600px]">
-		<Dialog.Header>
-			<Dialog.Title>{m['eventInvitationsAdmin.editDialogTitle']()}</Dialog.Title>
-			<Dialog.Description>
-				{#if editingInvitation}
-					{#if editingType === 'registered' && 'user' in editingInvitation}
-						{m['eventInvitationsAdmin.editDialogDescriptionUser']({
-							userName: getUserDisplayName(
-								editingInvitation.user,
-								m['eventInvitationsAdmin.unknownUser']()
-							)
-						})}
-					{:else if editingType === 'pending' && 'email' in editingInvitation}
-						{m['eventInvitationsAdmin.editDialogDescriptionEmail']({
-							email: editingInvitation.email
-						})}
-					{/if}
-				{/if}
-			</Dialog.Description>
-		</Dialog.Header>
-
-		<form
-			method="POST"
-			action="?/updateInvitation"
-			use:enhance={() => {
-				return async ({ update }) => {
-					await update();
-					resetEditForm();
-					showEditDialog = false;
-				};
-			}}
-			class="flex min-h-0 flex-1 flex-col gap-4"
-		>
-			<div class="min-h-0 flex-1 space-y-4 overflow-y-auto pr-1">
-				<input type="hidden" name="tier_ids" value={JSON.stringify(editFormData.tier_ids)} />
-
-				{#if editingInvitation}
-					<!-- Hidden email field -->
-					{#if editingType === 'registered' && 'user' in editingInvitation}
-						<input type="hidden" name="email" value={editingInvitation.user.email || ''} />
-					{:else if editingType === 'pending' && 'email' in editingInvitation}
-						<input type="hidden" name="email" value={editingInvitation.email} />
-					{/if}
-
-					<!-- Custom message -->
-					<div>
-						<label for="edit_custom_message" class="block text-sm font-medium">
-							{m['eventInvitationsAdmin.customMessageLabel']()}
-						</label>
-						<textarea
-							id="edit_custom_message"
-							name="custom_message"
-							bind:value={editFormData.custom_message}
-							placeholder={m['eventInvitationsAdmin.customMessagePlaceholder']()}
-							rows="3"
-							maxlength="500"
-							class="mt-1 w-full rounded-md border-2 border-gray-300 bg-white px-3 py-2 text-sm transition-colors placeholder:text-muted-foreground focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary focus:ring-offset-2 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100"
-						></textarea>
-						<p class="mt-1 text-xs text-muted-foreground">
-							{m['eventInvitationsAdmin.charactersCount']({
-								count: editFormData.custom_message.length
-							})}
-						</p>
-					</div>
-
-					<!-- Invitation properties -->
-					<div class="space-y-3 rounded-lg border border-border p-4">
-						<h4 class="text-sm font-semibold">{m['eventInvitationsAdmin.propertiesTitle']()}</h4>
-
-						<label class="flex items-center gap-2">
-							<input
-								type="checkbox"
-								name="waives_questionnaire"
-								value="true"
-								bind:checked={editFormData.waives_questionnaire}
-								class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-							/>
-							<span class="text-sm">{m['eventInvitationsAdmin.waiveQuestionnaire']()}</span>
-						</label>
-
-						<label class="flex items-center gap-2">
-							<input
-								type="checkbox"
-								name="waives_purchase"
-								value="true"
-								bind:checked={editFormData.waives_purchase}
-								class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-							/>
-							<span class="text-sm">{m['eventInvitationsAdmin.waivePurchase']()}</span>
-						</label>
-
-						<label class="flex items-center gap-2">
-							<input
-								type="checkbox"
-								name="waives_membership_required"
-								value="true"
-								bind:checked={editFormData.waives_membership_required}
-								class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-							/>
-							<span class="text-sm">{m['eventInvitationsAdmin.waiveMembership']()}</span>
-						</label>
-
-						<label class="flex items-center gap-2">
-							<input
-								type="checkbox"
-								name="waives_rsvp_deadline"
-								value="true"
-								bind:checked={editFormData.waives_rsvp_deadline}
-								class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-							/>
-							<span class="text-sm">{m['eventInvitationsAdmin.waiveDeadline']()}</span>
-						</label>
-
-						<label class="flex items-center gap-2">
-							<input
-								type="checkbox"
-								name="overrides_max_attendees"
-								value="true"
-								bind:checked={editFormData.overrides_max_attendees}
-								class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-							/>
-							<span class="text-sm">{m['eventInvitationsAdmin.overrideMaxAttendees']()}</span>
-						</label>
-					</div>
-
-					<!-- Ticket tier selection -->
-					{@render tierCheckboxes(editFormData.tier_ids, (ids) => (editFormData.tier_ids = ids))}
-				{/if}
-			</div>
-
-			<Dialog.Footer>
-				<Button type="button" variant="outline" onclick={() => (showEditDialog = false)}>
-					{m['eventInvitationsAdmin.cancel']()}
-				</Button>
-				<Button type="submit">
-					<Edit class="mr-2 h-4 w-4" aria-hidden="true" />
-					{m['eventInvitationsAdmin.updateInvitation']()}
-				</Button>
-			</Dialog.Footer>
-		</form>
-	</Dialog.Content>
-</Dialog.Root>
-
-<!-- Bulk Edit Dialog -->
-<Dialog.Root
-	open={showBulkEditDialog}
-	onOpenChange={(open) => {
-		showBulkEditDialog = open;
-		if (!open) resetBulkEditForm();
-	}}
->
-	<Dialog.Content class="flex max-h-[90dvh] flex-col sm:max-w-[600px]">
-		<Dialog.Header>
-			<Dialog.Title>{m['eventInvitationsAdmin.bulkEditDialogTitle']()}</Dialog.Title>
-			<Dialog.Description>
-				{m['eventInvitationsAdmin.bulkEditDialogDescription']({
-					count: totalSelected,
-					plural: totalSelected === 1 ? '' : 's'
-				})}
-			</Dialog.Description>
-		</Dialog.Header>
-
-		<form
-			method="POST"
-			action="?/bulkUpdateInvitations"
-			use:enhance={() => {
-				return async ({ update }) => {
-					await update();
-					resetBulkEditForm();
-					showBulkEditDialog = false;
-					clearSelections();
-				};
-			}}
-			class="flex min-h-0 flex-1 flex-col gap-4"
-		>
-			<div class="min-h-0 flex-1 space-y-4 overflow-y-auto pr-1">
-				<!-- Hidden emails field (JSON array) -->
-				<input
-					type="hidden"
-					name="emails"
-					value={JSON.stringify([
-						...Array.from(selectedRegisteredIds)
-							.map((id) => registeredInvitations.find((inv) => inv.id === id)?.user?.email)
-							.filter((e) => e),
-						...Array.from(selectedPendingIds)
-							.map((id) => pendingInvitations.find((inv) => inv.id === id)?.email)
-							.filter((e) => e)
-					])}
-				/>
-				<input type="hidden" name="tier_ids" value={JSON.stringify(bulkEditFormData.tier_ids)} />
-
-				<!-- Custom message -->
-				<div>
-					<label for="bulk_custom_message" class="block text-sm font-medium">
-						{m['eventInvitationsAdmin.customMessageLabel']()}
-					</label>
-					<textarea
-						id="bulk_custom_message"
-						name="custom_message"
-						bind:value={bulkEditFormData.custom_message}
-						placeholder={m['eventInvitationsAdmin.customMessagePlaceholder']()}
-						rows="3"
-						maxlength="500"
-						class="mt-1 w-full rounded-md border-2 border-gray-300 bg-white px-3 py-2 text-sm transition-colors placeholder:text-muted-foreground focus:border-primary focus:outline-none focus:ring-2 focus:ring-primary focus:ring-offset-2 dark:border-gray-600 dark:bg-gray-800 dark:text-gray-100"
-					></textarea>
-					<p class="mt-1 text-xs text-muted-foreground">
-						{m['eventInvitationsAdmin.charactersCount']({
-							count: bulkEditFormData.custom_message.length
-						})}
-					</p>
-				</div>
-
-				<!-- Invitation properties -->
-				<div class="space-y-3 rounded-lg border border-border p-4">
-					<h4 class="text-sm font-semibold">{m['eventInvitationsAdmin.propertiesTitle']()}</h4>
-
-					<label class="flex items-center gap-2">
-						<input
-							type="checkbox"
-							name="waives_questionnaire"
-							value="true"
-							bind:checked={bulkEditFormData.waives_questionnaire}
-							class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-						/>
-						<span class="text-sm">{m['eventInvitationsAdmin.waiveQuestionnaire']()}</span>
-					</label>
-
-					<label class="flex items-center gap-2">
-						<input
-							type="checkbox"
-							name="waives_purchase"
-							value="true"
-							bind:checked={bulkEditFormData.waives_purchase}
-							class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-						/>
-						<span class="text-sm">{m['eventInvitationsAdmin.waivePurchase']()}</span>
-					</label>
-
-					<label class="flex items-center gap-2">
-						<input
-							type="checkbox"
-							name="waives_membership_required"
-							value="true"
-							bind:checked={bulkEditFormData.waives_membership_required}
-							class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-						/>
-						<span class="text-sm">{m['eventInvitationsAdmin.waiveMembership']()}</span>
-					</label>
-
-					<label class="flex items-center gap-2">
-						<input
-							type="checkbox"
-							name="waives_rsvp_deadline"
-							value="true"
-							bind:checked={bulkEditFormData.waives_rsvp_deadline}
-							class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-						/>
-						<span class="text-sm">{m['eventInvitationsAdmin.waiveDeadline']()}</span>
-					</label>
-
-					<label class="flex items-center gap-2">
-						<input
-							type="checkbox"
-							name="overrides_max_attendees"
-							value="true"
-							bind:checked={bulkEditFormData.overrides_max_attendees}
-							class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-						/>
-						<span class="text-sm">{m['eventInvitationsAdmin.overrideMaxAttendees']()}</span>
-					</label>
-				</div>
-
-				<!-- Ticket tier selection -->
-				{@render tierCheckboxes(
-					bulkEditFormData.tier_ids,
-					(ids) => (bulkEditFormData.tier_ids = ids)
-				)}
-			</div>
-
-			<Dialog.Footer>
-				<Button type="button" variant="outline" onclick={() => (showBulkEditDialog = false)}>
-					{m['eventInvitationsAdmin.cancel']()}
-				</Button>
-				<Button type="submit">
-					<Edit class="mr-2 h-4 w-4" aria-hidden="true" />
-					{m['eventInvitationsAdmin.updateInvitations']({
-						count: totalSelected,
-						plural: totalSelected === 1 ? '' : 's'
-					})}
-				</Button>
-			</Dialog.Footer>
-		</form>
-	</Dialog.Content>
-</Dialog.Root>
-
-{#snippet invitationProperties(
-	invitation: EventInvitationListSchema | PendingEventInvitationListSchema
-)}
-	<div class="flex flex-wrap gap-1">
-		{#if invitation.waives_questionnaire}
-			<span
-				class="inline-flex items-center rounded-full bg-blue-100 px-2 py-0.5 text-xs font-medium text-blue-800 dark:bg-blue-900 dark:text-blue-200"
-				title={m['invitationListTab.waivesQuestionnaireTitle']()}
-			>
-				{m['eventInvitationsAdmin.noQuestionnaire']()}
-			</span>
-		{/if}
-		{#if invitation.waives_purchase}
-			<span
-				class="inline-flex items-center rounded-full bg-green-100 px-2 py-0.5 text-xs font-medium text-green-800 dark:bg-green-900 dark:text-green-200"
-				title={m['invitationListTab.waivesPurchaseTitle']()}
-			>
-				{m['eventInvitationsAdmin.free']()}
-			</span>
-		{/if}
-		{#if invitation.waives_membership_required}
-			<span
-				class="inline-flex items-center rounded-full bg-purple-100 px-2 py-0.5 text-xs font-medium text-purple-800 dark:bg-purple-900 dark:text-purple-200"
-				title={m['invitationListTab.waivesMembershipTitle']()}
-			>
-				{m['eventInvitationsAdmin.noMembership']()}
-			</span>
-		{/if}
-		{#if invitation.waives_rsvp_deadline}
-			<span
-				class="inline-flex items-center rounded-full bg-orange-100 px-2 py-0.5 text-xs font-medium text-orange-800 dark:bg-orange-900 dark:text-orange-200"
-				title={m['invitationListTab.waivesRsvpDeadlineTitle']()}
-			>
-				{m['eventInvitationsAdmin.noDeadline']()}
-			</span>
-		{/if}
-		{#if invitation.overrides_max_attendees}
-			<span
-				class="inline-flex items-center rounded-full bg-red-100 px-2 py-0.5 text-xs font-medium text-red-800 dark:bg-red-900 dark:text-red-200"
-				title={m['invitationListTab.overridesMaxAttendeesTitle']()}
-			>
-				{m['eventInvitationsAdmin.overrideCap']()}
-			</span>
-		{/if}
-		{#if invitation.tiers?.length}
-			{#each invitation.tiers as tier (tier.id)}
-				<span
-					class="inline-flex items-center rounded-full bg-indigo-100 px-2 py-0.5 text-xs font-medium text-indigo-800 dark:bg-indigo-900 dark:text-indigo-200"
-					title={m['invitationListTab.assignedTierTitle']({ name: tier.name })}
-				>
-					{tier.name}
-				</span>
-			{/each}
-		{/if}
-		{#if invitation.custom_message}
-			<span
-				class="inline-flex items-center rounded-full bg-gray-100 px-2 py-0.5 text-xs font-medium text-gray-800 dark:bg-gray-900 dark:text-gray-200"
-				title={invitation.custom_message}
-			>
-				Has {m['eventInvitationsAdmin.headerMessage']()}
-			</span>
-		{/if}
-	</div>
+{#snippet registeredHeaders()}
+	<th
+		class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
+	>
+		{m['eventInvitationsAdmin.headerUser']()}
+	</th>
+	<th
+		class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
+	>
+		{m['eventInvitationsAdmin.headerEmail']()}
+	</th>
 {/snippet}
 
-{#snippet invitationPropertiesCheckboxes()}
-	<div class="space-y-3 rounded-lg border border-border p-4">
-		<h4 class="text-sm font-semibold">{m['eventInvitationsAdmin.propertiesTitle']()}</h4>
-
-		<label class="flex items-center gap-2">
-			<input
-				type="checkbox"
-				name="waives_questionnaire"
-				value="true"
-				class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
+{#snippet registeredCells(invitation: EventInvitationListSchema)}
+	<!-- User -->
+	<td class="px-4 py-4">
+		<div class="flex items-center gap-2">
+			<UserAvatar
+				profilePictureUrl={invitation.user.profile_picture_url}
+				previewUrl={invitation.user.profile_picture_preview_url}
+				thumbnailUrl={invitation.user.profile_picture_thumbnail_url}
+				displayName={getUserDisplayName(invitation.user, m['eventInvitationsAdmin.unknownUser']())}
+				firstName={invitation.user.first_name}
+				lastName={invitation.user.last_name}
+				size="sm"
+				clickable={true}
 			/>
-			<span class="text-sm">{m['eventInvitationsAdmin.waiveQuestionnaire']()}</span>
-		</label>
-
-		<label class="flex items-center gap-2">
-			<input
-				type="checkbox"
-				name="waives_purchase"
-				value="true"
-				class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-			/>
-			<span class="text-sm">{m['eventInvitationsAdmin.waivePurchase']()}</span>
-		</label>
-
-		<label class="flex items-center gap-2">
-			<input
-				type="checkbox"
-				name="waives_membership_required"
-				value="true"
-				class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-			/>
-			<span class="text-sm">{m['eventInvitationsAdmin.waiveMembership']()}</span>
-		</label>
-
-		<label class="flex items-center gap-2">
-			<input
-				type="checkbox"
-				name="waives_rsvp_deadline"
-				value="true"
-				class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-			/>
-			<span class="text-sm">{m['eventInvitationsAdmin.waiveDeadline']()}</span>
-		</label>
-
-		<label class="flex items-center gap-2">
-			<input
-				type="checkbox"
-				name="overrides_max_attendees"
-				value="true"
-				class="h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary focus:ring-offset-2"
-			/>
-			<span class="text-sm">{m['eventInvitationsAdmin.overrideMaxAttendees']()}</span>
-		</label>
-	</div>
-{/snippet}
-
-{#snippet tierCheckboxes(selectedIds: string[], onUpdate: (ids: string[]) => void)}
-	{#if ticketTiers.length > 0}
-		<div class="space-y-3 rounded-lg border border-border p-4">
-			<h4 class="text-sm font-semibold">{m['invitationListTab.assignTicketTiersTitle']()}</h4>
-			<p class="text-xs text-muted-foreground">
-				{m['invitationListTab.assignTicketTiersDescription']()}
-			</p>
-			{#each ticketTiers as tier (tier.id ?? tier.name)}
-				{#if tier.id}
-					<label class="flex cursor-pointer items-start gap-2">
-						<input
-							type="checkbox"
-							checked={selectedIds.includes(tier.id)}
-							onchange={(e) => {
-								const checked = e.currentTarget.checked;
-								if (checked && tier.id && !selectedIds.includes(tier.id)) {
-									onUpdate([...selectedIds, tier.id]);
-								} else if (!checked && tier.id) {
-									onUpdate(selectedIds.filter((id) => id !== tier.id));
-								}
-							}}
-							class="mt-0.5 h-4 w-4 rounded border-gray-300 text-primary focus:ring-2 focus:ring-primary"
-						/>
-						<div class="flex-1">
-							<span class="text-sm font-medium">{tier.name}</span>
-						</div>
-					</label>
-				{/if}
-			{/each}
+			<span class="text-sm font-medium">
+				{getUserDisplayName(invitation.user, m['eventInvitationsAdmin.unknownUser']())}
+			</span>
 		</div>
-	{/if}
+	</td>
+
+	<!-- Email -->
+	<td class="px-4 py-4 text-sm text-muted-foreground">
+		{invitation.user.email || 'N/A'}
+	</td>
+{/snippet}
+
+{#snippet registeredEmpty()}
+	<UserPlus class="mx-auto mb-2 h-8 w-8 text-muted-foreground" aria-hidden="true" />
+	<p class="text-sm text-muted-foreground">
+		{m['eventInvitationsAdmin.noRegisteredInvitations']()}
+	</p>
+{/snippet}
+
+{#snippet pendingHeaders()}
+	<th
+		class="px-4 py-3 text-left text-xs font-medium uppercase tracking-wider text-muted-foreground"
+	>
+		{m['eventInvitationsAdmin.headerEmail']()}
+	</th>
+{/snippet}
+
+{#snippet pendingCells(invitation: PendingEventInvitationListSchema)}
+	<!-- Email -->
+	<td class="px-4 py-4 text-sm font-medium">{invitation.email}</td>
+{/snippet}
+
+{#snippet pendingEmpty()}
+	<Mail class="mx-auto mb-2 h-8 w-8 text-muted-foreground" aria-hidden="true" />
+	<p class="text-sm text-muted-foreground">
+		{m['eventInvitationsAdmin.noPendingInvitations']()}
+	</p>
 {/snippet}
